requirements.txt
PyJWT==1.7.1
requests-oauthlib==1.2.0
